19/1/2025 update User and Role

+++ src/main/java/com/example/bookverse/service/impl/RoleServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.bookverse.service.impl;
 
 import com.example.bookverse.domain.Role;
+import com.example.bookverse.exception.global.IdInvalidException;
+import com.example.bookverse.exception.role.ExistRoleNameException;
 import com.example.bookverse.repository.RoleRepository;
 import com.example.bookverse.service.RoleService;
 import org.springframework.stereotype.Service;
@@ -17,17 +19,32 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Role create(Role role) {
+    public Role create(Role role) throws ExistRoleNameException {
+        if (this.roleRepository.existsByName(role.getName())) {
+            throw new ExistRoleNameException(role.getName() + " already exist");
+        }
         return this.roleRepository.save(role);
     }
 
     @Override
-    public Role update(Role role) {
-        return this.roleRepository.save(role);
+    public Role update(Role role) throws IdInvalidException {
+        Role roleInDB = this.roleRepository.findById(role.getId()).orElse(null);
+        if (roleInDB == null) {
+            throw new IdInvalidException("Role not found");
+        }
+        else {
+            if (role.getDescription() != null && !role.getDescription().equals(roleInDB.getDescription())) {
+                roleInDB.setDescription(role.getDescription());
+            }
+            return roleRepository.save(roleInDB);
+        }
     }
 
     @Override
-    public Role fetchRoleById(long id) {
+    public Role fetchRoleById(long id) throws IdInvalidException {
+        if (!this.roleRepository.existsById(id)) {
+            throw new IdInvalidException("Role not found");
+        }
         return this.roleRepository.findById(id).orElse(null);
     }
 
@@ -37,7 +54,10 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public void delete(long id) {
+    public void delete(long id) throws IdInvalidException {
+        if (!this.roleRepository.existsById(id)) {
+            throw new IdInvalidException("Role not found");
+        }
         this.roleRepository.deleteById(id);
     }
 }
